ci: migrate to github actions
To migrate to GitHub Actions and our library workflows a few changes are
needed:

First up, removing the .travis.yml and adding in the
.github/workflows/dhis2-verify-lib.yml.

The lint job is left commented out, as there are 600+ linting errors
reported (all autofixable it seems), but fixing the lint errors should
be in a different commit, as this needs to be backported and is designed
to be easy to move to other branches.

The main entrypoint in package.json was pointing to a directory that
doesn't exist, so that has been updated to build/index.js.

When moving to the d2-utils-release script, we will no longer publish
this as a flat package, so this is potentially a breaking change. I
think that it worked because the lib was compiled into build/index.js,
and then since the main entrypoint in package.json did not exist, the
file-loader fell back to loading the index.js from the package, which
happened to exist in the root folder because of the flatpack structure
in the released artifact.

This should "just work" with the correct entry-point even if we no
longer release the flat package.

Old version of node-sass has been bumped to 4.13.1 which is close to the
previous to not have any breaking changes, but supports node versions >=
10.

Fix: https://jira.dhis2.org/browse/TECH-512